fixing bugs
Former-commit-id: 4fc9d9621a9ae8ec7f44a26cd58f8ce46af2e3f0

+++ Kinvey/Kinvey/FileStore.swift
@@ -142,7 +142,7 @@ public class FileStore {
                 } else if let fromFile = fromFile,
                     let path = fromFile.path,
                     let attrs = try? NSFileManager.defaultManager().attributesOfItemAtPath((path as NSString).stringByExpandingTildeInPath),
-                    let fileSize = attrs[NSFileSize] as? UInt64
+                    let fileSize = attrs[NSFileSize] as? UInt
                 {
                     request.setValue("bytes */\(fileSize)", forHTTPHeaderField: "Content-Range")
                 }
